// standing-out/StandingOutStore/AngularSource/app/models/session-attendee.ts
export class SessionAttendee {
    sessionAttendeeId: string;
    userId: string;
    email: string;
    firstName: string;
    lastName: string;
    createdDate: Date;
    refunded: boolean;
    removed: boolean;
    isDeleted: boolean;
    isRefundStudentInitiated: boolean;
    fullName: string;
    status: string;
}